Fix response interpretation according to specification
The specification was changed a little.

Status code 404 now really means what it means, which is that the
service doesn't recognize the requested URL. This would mean an
incorrectly configured GDPR API URL. Since previously 404 was
interpreted in a different way, informational text is left into the tool
so that GDPR API implementations following the old specification get
appropriate help.

The GET operation accepts a 204 status code, which indicates that the
service doesn't contain any data for the profile.