Add an issuer type configuration option
The behaviour of the simulated authorization server can be changed based
on this setting.

Added a dedicated enum type for this setting since the allowed values
are limited to two options. Since the `AppConfig`'s `set_config` method
creates a value of the proper type, it raises an exception if an invalid
value is provided.