Generate access tokens based on issuer type
The scopes in the generated access tokens are different in Tunnistamo
and Keycloak. The existing style was the Tunnistamo style. Keycloak
style is added in this commit.

The `GDPR_API_QUERY_SCOPE` and `GDPR_API_DELETE_SCOPE` settings are used
for both styles. But for Keycloak style only the part after the last dot
is used. This makes it possible to for example keep the value as
`exampleapi.gdprquery`; for Tunnistamo the value is used as is and for
Keycloak only `gdprquery` part is used, leading to correct looking
access tokens in all cases.